+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.2.0 - YYYY-MM-DD #
 
-- Automatic paging.
+- Plugin "stdin".
 
 # 0.1.0 - 2016-02-13 #