+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# 0.2.0 - YYYY-MM-DD #
+# 0.3.0 - YYYY-MM-DD #
+
+- Plugin "pager": automatic paging.
+
+# 0.2.0 - 2016-02-20 #
 
 - Disable plugin option.
 
 # 0.1.0 - 2016-02-13 #
 
-- Plugin "ls".
+- Plugin "ls": POSIX `ls` command.